+++ xenon-common/src/main/java/com/vmware/xenon/services/common/LuceneDocumentIndexService.java
@@ -731,12 +731,12 @@ public class LuceneDocumentIndexService extends StatelessService {
     public void handleGetImpl(Operation get) throws Throwable {
         String selfLink = null;
         Long version = null;
-        ServiceOption targetIndex = ServiceOption.NONE;
+        ServiceOption serviceOption = ServiceOption.NONE;
 
         EnumSet<QueryOption> options = EnumSet.noneOf(QueryOption.class);
         if (get.hasPragmaDirective(Operation.PRAGMA_DIRECTIVE_INDEX_CHECK)) {
             // fast path for checking if a service exists, and loading its latest state
-            targetIndex = ServiceOption.PERSISTENCE;
+            serviceOption = ServiceOption.PERSISTENCE;
             // the GET operation URI is set to the service we want to load, not the self link
             // of the index service. This is only possible when the operation was directly
             // dispatched from the local host, on the index service
@@ -748,7 +748,12 @@ public class LuceneDocumentIndexService extends StatelessService {
             String cap = params.get(UriUtils.URI_PARAM_CAPABILITY);
 
             if (cap != null) {
-                targetIndex = ServiceOption.valueOf(cap);
+                serviceOption = ServiceOption.valueOf(cap);
+            }
+
+            if (serviceOption == ServiceOption.IMMUTABLE) {
+                options.add(QueryOption.INCLUDE_ALL_VERSIONS);
+                serviceOption = ServiceOption.PERSISTENCE;
             }
 
             if (params.containsKey(UriUtils.URI_PARAM_INCLUDE_DELETED)) {
@@ -796,7 +801,7 @@ public class LuceneDocumentIndexService extends StatelessService {
             return;
         }
 
-        if (targetIndex == ServiceOption.PERSISTENCE) {
+        if (serviceOption == ServiceOption.PERSISTENCE) {
             // specific index requested but no results, return empty response
             get.setBodyNoCloning(rsp).complete();
             return;
@@ -1403,8 +1408,7 @@ public class LuceneDocumentIndexService extends StatelessService {
         // Keep duplicates out
         Set<String> uniques = new LinkedHashSet<>(rsp.documentLinks);
         final boolean hasCountOption = options.contains(QueryOption.COUNT);
-        final boolean hasIncludeAllVersionsOption = options
-                .contains(QueryOption.INCLUDE_ALL_VERSIONS);
+        boolean hasIncludeAllVersionsOption = options.contains(QueryOption.INCLUDE_ALL_VERSIONS);
         Set<String> linkWhiteList = null;
         if (qs != null && qs.context != null && qs.context.documentLinkWhiteList != null) {
             linkWhiteList = qs.context.documentLinkWhiteList;
@@ -2229,6 +2233,7 @@ public class LuceneDocumentIndexService extends StatelessService {
         }
 
         long startNanos = System.nanoTime();
+
         wr.addDocument(doc);
         long durationNanos = System.nanoTime() - startNanos;
         setTimeSeriesStat(STAT_NAME_INDEXED_DOCUMENT_COUNT, AGGREGATION_TYPE_SUM, 1);